Move functions back
