+++ lib/plugins/block_actions.js
@@ -1,6 +1,18 @@
 const { Vec3 } = require('vec3')
 
-const { CARDINAL_OFFSETS, CHEST_FACING_MAP, getChestType, parseChestMetadata } = require('../blocks.js')
+const CARDINAL_OFFSETS = {
+  north: new Vec3(0, 0, -1),
+  south: new Vec3(0, 0, 1),
+  west: new Vec3(-1, 0, 0),
+  east: new Vec3(1, 0, 0)
+}
+
+const CHEST_FACING_MAP = {
+  north: { west: 'right', east: 'left' },
+  south: { west: 'left', east: 'right' },
+  west: { north: 'left', south: 'right' },
+  east: { north: 'right', south: 'left' }
+}
 
 module.exports = inject
 
@@ -9,6 +21,39 @@ function inject (bot) {
 
   // Stores how many players have currently open a container at a certain position
   const openCountByPos = {}
+
+  bot.parseChestMetadata = (chestBlock) => {
+    const chestTypes = ['single', 'right', 'left']
+
+    return bot.supportFeature('doesntHaveChestType')
+      ? { facing: Object.keys(CARDINAL_OFFSETS)[chestBlock.metadata - 2] }
+      : {
+          waterlogged: !(chestBlock.metadata & 1),
+          type: chestTypes[(chestBlock.metadata >> 1) % 3],
+          facing: Object.keys(CARDINAL_OFFSETS)[Math.floor(chestBlock.metadata / 6)]
+        }
+  }
+
+  bot.getChestType = (chestBlock) => { // Returns 'single', 'right' or 'left'
+    if (bot.supportFeature('doesntHaveChestType')) {
+      const facing = bot.parseChestMetadata(chestBlock).facing
+
+      if (!facing) return 'single'
+
+      // We have to check if the adjacent blocks in the perpendicular cardinals are the same type
+      const perpendicularCardinals = Object.keys(CHEST_FACING_MAP[facing])
+      for (const cardinal of perpendicularCardinals) {
+        const cardinalOffset = CARDINAL_OFFSETS[cardinal]
+        if (bot.blockAt(chestBlock.position.plus(cardinalOffset))?.type === chestBlock.type) {
+          return CHEST_FACING_MAP[cardinal][facing]
+        }
+      }
+
+      return 'single'
+    } else {
+      return bot.parseChestMetadata(chestBlock).type
+    }
+  }
 
   bot._client.on('block_action', (packet) => {
     const pt = new Vec3(packet.location.x, packet.location.y, packet.location.z)
@@ -29,10 +74,10 @@ function inject (bot) {
       let block2 = null
 
       if (blockName === 'chest' || blockName === 'trapped_chest') {
-        const chestType = getChestType(bot, block)
+        const chestType = bot.getChestType(block)
         if (chestType === 'right') {
-          const index = Object.values(CHEST_FACING_MAP[parseChestMetadata(bot, block).facing]).indexOf('left')
-          const cardinalBlock2 = Object.keys(CHEST_FACING_MAP[parseChestMetadata(bot, block).facing])[index]
+          const index = Object.values(CHEST_FACING_MAP[bot.parseChestMetadata(block).facing]).indexOf('left')
+          const cardinalBlock2 = Object.keys(CHEST_FACING_MAP[bot.parseChestMetadata(block).facing])[index]
           const block2Position = block.position.plus(CARDINAL_OFFSETS[cardinalBlock2])
           block2 = bot.blockAt(block2Position)
         } else if (chestType === 'left') return // Omit left part of the chest so 'chestLidMove' doesn't emit twice when it's a double chest
