+++ lib/plugins/blocks.js
@@ -5,8 +5,6 @@ const { onceWithCleanup } = require('../promise_utils')
 
 const { OctahedronIterator } = require('prismarine-world').iterators
 
-const { getChestType } = require('../blocks.js')
-
 module.exports = inject
 
 const paintingFaceToVec = [
@@ -210,7 +208,7 @@ function inject (bot, { version, storageBuilder, hideErrors }) {
       if (block.type !== chestBlockId && block.type !== trappedChestBlockId) return false
 
       // Only match single chests and the right side of double chests
-      const chestType = getChestType(bot, block)
+      const chestType = bot.getChestType(block)
       if (chestType === 'single' || chestType === 'right') return true
     }
